# БР2.2/Свинин Артем/homeworks/hw2/openapi.yaml
openapi: 3.0.0
info:
  title: Rental Service API
  version: 0.0.0
...
paths:
...
  /favourites:
    post:
      operationId: FavouritesApi_add
      description: Добавить объект в избранное (userId берется из JWT)
      parameters: []
      responses:
        '201':
          description: Ресурс успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favourite'
        '409':
          description: Конфликт - ресурс уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Favourites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                propertyId:
                  type: integer
                  format: int32
              required:
                - propertyId
    get:
      operationId: FavouritesApi_list
      description: Получить список избранных объектов текущего пользователя
      parameters:
        - $ref: '#/components/parameters/PaginationParams.page'
        - $ref: '#/components/parameters/PaginationParams.pageSize'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - total
                  - page
                  - pageSize
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Favourite'
                  total:
                    type: integer
                    format: int32
                  page:
                    type: integer
                    format: int32
                  pageSize:
                    type: integer
                    format: int32
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Favourites
...
  /properties:
    get:
      operationId: PropertiesApi_list
      description: Поиск объектов недвижимости с фильтрацией и пагинацией
      parameters:
        - $ref: '#/components/parameters/PaginationParams.page'
        - $ref: '#/components/parameters/PaginationParams.pageSize'
        - name: search
          in: query
          required: false
          description: Поиск по названию или описанию
          schema:
            type: string
          explode: false
        - name: minPricePerDay
          in: query
          required: false
          schema:
            type: number
            format: decimal
          explode: false
        - name: maxPricePerDay
          in: query
          required: false
          schema:
            type: number
            format: decimal
          explode: false
        - name: minPricePerMonth
          in: query
          required: false
          schema:
            type: number
            format: decimal
          explode: false
        - name: maxPricePerMonth
          in: query
          required: false
          schema:
            type: number
            format: decimal
          explode: false
        - name: minArea
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: maxArea
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: bedrooms
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: bathrooms
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: maxGuests
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: facilities
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/FacilityType'
          explode: false
        - name: country
          in: query
          required: false
          description: Поиск по стране
          schema:
            type: string
          explode: false
        - name: city
          in: query
          required: false
          description: Поиск по городу
          schema:
            type: string
          explode: false
        - name: district
          in: query
          required: false
          description: Поиск по району
          schema:
            type: string
          explode: false
        - name: address
          in: query
          required: false
          description: Поиск по полному адресу (подстрока)
          schema:
            type: string
          explode: false
        - name: isAvailable
          in: query
          required: false
          schema:
            type: boolean
          explode: false
        - name: typeId
          in: query
          required: false
          description: Фильтрация по типу недвижимости
          schema:
            type: integer
            format: int32
          explode: false
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - total
                  - page
                  - pageSize
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Property'
                  total:
                    type: integer
                    format: int32
                  page:
                    type: integer
                    format: int32
                  pageSize:
                    type: integer
                    format: int32
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Properties
    post:
      operationId: PropertiesApi_create
      description: Создать новый объект недвижимости
      parameters: []
      responses:
        '201':
          description: Ресурс успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Properties
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                typeId:
                  type: integer
                  format: int32
                  description: ownerId устанавливается из JWT токена
                addressId:
                  type: integer
                  format: int32
                title:
                  type: string
                description:
                  type: string
                pricePerDay:
                  type: number
                  format: decimal
                  minimum: 0
                pricePerMonth:
                  type: number
                  format: decimal
                  minimum: 0
                areaSqM:
                  type: integer
                  format: int32
                  minimum: 1
                maxGuests:
                  type: integer
                  format: int32
                  minimum: 1
                bedrooms:
                  type: integer
                  format: int32
                  minimum: 0
                bathrooms:
                  type: integer
                  format: int32
                  minimum: 0
                isAvailable:
                  type: boolean
                minRentDays:
                  type: integer
                  format: int32
                  minimum: 1
                maxRentDays:
                  type: integer
                  format: int32
              required:
                - addressId
                - title
                - description
                - pricePerDay
                - pricePerMonth
                - areaSqM
                - maxGuests
                - bedrooms
                - bathrooms
                - minRentDays
...
  /users:
    get:
      operationId: UsersApi_list
      description: Получить список пользователей с поиском и фильтрацией
      parameters:
        - $ref: '#/components/parameters/PaginationParams.page'
        - $ref: '#/components/parameters/PaginationParams.pageSize'
        - name: search
          in: query
          required: false
          description: Поиск по ФИО (имя, фамилия или отчество)
          schema:
            type: string
          explode: false
        - name: role
          in: query
          required: false
          description: Фильтрация по роли
          schema:
            $ref: '#/components/schemas/Role'
          explode: false
        - name: isActive
          in: query
          required: false
          description: Фильтрация по статусу активности
          schema:
            type: boolean
          explode: false
        - name: isVerified
          in: query
          required: false
          description: Фильтрация по статусу верификации
          schema:
            type: boolean
          explode: false
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - total
                  - page
                  - pageSize
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
                  total:
                    type: integer
                    format: int32
                  page:
                    type: integer
                    format: int32
                  pageSize:
                    type: integer
                    format: int32
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      tags:
        - Users
...
security:
  - BearerAuth: []
components:
  parameters:
    PaginationParams.page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        format: int32
      explode: false
    PaginationParams.pageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        format: int32
      explode: false
  schemas:
    Address:
      type: object
      required:
        - id
        - country
        - city
        - district
        - houseNumber
      properties:
        id:
          type: integer
          format: int32
        country:
          type: string
        city:
          type: string
        district:
          type: string
        houseNumber:
          type: string
        apartment:
          type: string
        postalCode:
          type: string
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
        fullAddress:
          type: string
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
    FacilityType:
      type: string
      enum:
        - WIFI
        - WASHING_MACHINE
        - AIR_CONDITIONING
        - PARKING
        - KITCHEN
        - TV
        - HEATING
        - ELEVATOR
        - POOL
        - GYM
    Favourite:
      type: object
      required:
        - id
        - userId
        - propertyId
        - addedAt
      properties:
        id:
          type: integer
          format: int32
        userId:
          type: integer
          format: int32
        propertyId:
          type: integer
          format: int32
        addedAt:
          type: string
          format: date-time
...
    Property:
      type: object
      required:
        - id
        - ownerId
        - addressId
        - title
        - description
        - pricePerDay
        - pricePerMonth
        - areaSqM
        - maxGuests
        - bedrooms
        - bathrooms
        - isAvailable
        - minRentDays
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int32
        ownerId:
          type: integer
          format: int32
          description: Устанавливается автоматически из JWT
        typeId:
          type: integer
          format: int32
        type:
          $ref: '#/components/schemas/PropertyType'
        addressId:
          type: integer
          format: int32
        address:
          $ref: '#/components/schemas/Address'
        title:
          type: string
        description:
          type: string
        pricePerDay:
          type: number
          format: decimal
          minimum: 0
        pricePerMonth:
          type: number
          format: decimal
          minimum: 0
        areaSqM:
          type: integer
          format: int32
          minimum: 1
        maxGuests:
          type: integer
          format: int32
          minimum: 1
        bedrooms:
          type: integer
          format: int32
          minimum: 0
        bathrooms:
          type: integer
          format: int32
          minimum: 0
        isAvailable:
          type: boolean
        minRentDays:
          type: integer
          format: int32
          minimum: 1
        maxRentDays:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
          description: Генерируется сервером
        updatedAt:
          type: string
          format: date-time
          description: Обновляется сервером
...
    PropertyType:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
...
    Role:
      type: string
      enum:
        - ADMIN
        - OWNER
        - OCCUPIER
...
    UserResponse:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - role
        - isVerified
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        isVerified:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer